python/requirements.txt
requests>=2.28.0
argparse>=1.4.0
anthropic>=0.21.0
pandas>=2.0.0
python-dotenv>=1.0.0